cte refactoring
